test: split runtime instance contracts
